requirements.txt
cassandra-driver==3.29.2
click==8.1.8
geomet==0.2.1.post1
pika==1.3.2
pika-stubs==0.1.3
python-dotenv==1.0.1
six==1.17.0
